Move renderActivePoints to separate file
